Bug: when import is on last line in input file, operation stops
It seems like when we get to the last line to be parsed having an import
reference, everything fails... silently. adding a new line to the end of the
file to be imported/parsed seems to correct this, but I'm not entirely clear
as to why at the moment.

Update cli placeholder.